+++ container-images/mds-collector-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-collector-api
 
+## 0.2.22
+
+### Patch Changes
+
+- @mds-core/mds-collector-api@0.1.94
+
 ## 0.2.21
 
 ### Patch Changes