+++ packages/mds-collector-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mds-core/mds-collector-api
 
+## 0.1.94
+
+### Patch Changes
+
+- @mds-core/mds-api-helpers@0.3.13
+
 ## 0.1.93
 
 ### Patch Changes