+++ packages/mds-policy-author-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mds-core/mds-policy-author-api
 
+## 0.8.21
+
+### Patch Changes
+
+- Updated dependencies [e43d6969]
+  - @mds-core/mds-policy-service@0.13.2
+  - @mds-core/mds-api-helpers@0.3.13
+
 ## 0.8.20
 
 ### Patch Changes